matter: merge upstream commits up to June 4th
Merge Matter upstream changes and update NCS Matter samples
accordingly.

+++ samples/matter/lock/src/zcl_callbacks.cpp
@@ -9,16 +9,16 @@
 #include "app_task.h"
 #include "bolt_lock_manager.h"
 
-#include "gen/attribute-id.h"
-#include "gen/cluster-id.h"
-#include "gen/command-id.h"
-#include <app/util/af.h>
+#include <app/common/gen/attribute-id.h>
+#include <app/common/gen/attribute-type.h>
+#include <app/common/gen/cluster-id.h>
 #include <app/util/af-types.h>
+#include <app/util/af.h>
 
 using namespace ::chip;
 
 void emberAfPostAttributeChangeCallback(EndpointId endpoint, ClusterId clusterId, AttributeId attributeId, uint8_t mask,
-					uint16_t manufacturerCode, uint8_t type, uint8_t size, uint8_t *value)
+					uint16_t manufacturerCode, uint8_t type, uint16_t size, uint8_t *value)
 {
 	if (clusterId != ZCL_ON_OFF_CLUSTER_ID) {
 		ChipLogProgress(Zcl, "Unknown cluster ID: %d", clusterId);
